# docs/api/openapi-v1.snapshot.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agentor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agent-runs": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Lists agent runs with stable ordering (newest first) and pagination.",
        "operationId": "ListAgentRuns",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Starts a new deterministic agent run. Optional Idempotency-Key replays identical payloads to the same run; mismatched payloads return 409.",
        "operationId": "StartAgentRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentRunRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent-runs/{runId}": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Returns an agent run by ID.",
        "operationId": "GetAgentRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent-runs/{runId}/trace": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Returns execution trace events for an agent run.",
        "operationId": "GetAgentRunTrace",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent-runs/{runId}/steps": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Returns steps for an agent run.",
        "operationId": "GetAgentRunSteps",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent-runs/{runId}/tool-calls": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Returns tool calls for an agent run (flattened across steps, ordered by step then occurrence).",
        "operationId": "GetAgentRunToolCalls",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent-runs/{runId}/manifest": {
      "get": {
        "tags": [
          "AgentRuns"
        ],
        "summary": "Returns the run manifest for a completed agent run.",
        "operationId": "GetRunManifest",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NullableOfRunExecutionMode": {
        "enum": [
          "LegacyFakeTool",
          "SingleTool",
          "Plan",
          "Recipe",
          "Skill",
          "ModelCall",
          "McpTool",
          "ExternalAgent",
          null
        ],
        "default": null,
        "nullable": true
      },
...
      "StartAgentRunRequestDto": {
        "required": [
          "agentName",
          "objective"
        ],
        "type": "object",
        "properties": {
          "agentName": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "default": null,
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "knowledgeScopeId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/NullableOfRunExecutionMode"
          },
          "recipeId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "planId": {
            "type": "string",
            "format": "uuid",
            "default": null,
            "nullable": true
          },
          "toolKey": {
            "type": "string",
            "default": null,
            "nullable": true
          },
          "skillKey": {
            "type": "string",
            "default": null,
            "nullable": true
          },
          "input": {
            "type": "object",
            "default": null,
            "nullable": true
          },
          "toolInputPayload": {
            "default": null
          }
        }
      },
...
    }
  },
...
}